+++ src/react/loom-app/app/hooks/use-source/index.ts
@@ -1,4 +1,4 @@
-import { CellType, SourceType } from "src/shared/loom-state/types/loom-state";
+import { SourceType } from "src/shared/loom-state/types/loom-state";
 import { useLogger } from "src/shared/logger";
 import { useLoomState } from "../../../loom-state-provider";
 import SourceAddCommand from "src/shared/loom-state/commands/source-add-command";
@@ -6,8 +6,8 @@ import SourceDeleteCommand from "src/shared/loom-state/commands/source-delete-co
 import findDataFromSources from "src/shared/loom-state/find-data-from-sources";
 import { useAppMount } from "src/react/loom-app/app-mount-provider";
 import React, { useCallback } from "react";
-import { FrontMatterType } from "./types";
-import CellNotFoundError from "src/shared/error/cell-not-found-error";
+import { FrontMatterType } from "../../../../../shared/deserialize-frontmatter/types";
+import { deserializeFrontmatterKeys } from "src/shared/deserialize-frontmatter";
 
 export const useSource = () => {
 	const logger = useLogger();
@@ -20,57 +20,10 @@ export const useSource = () => {
 
 	React.useEffect(() => {
 		const { columns, rows } = loomState.model;
-		setFrontMatterKeys((prevState) => {
-			const newFrontmatterKeys = new Map(prevState);
-
-			const sourceFileColumn = columns.find(
-				(column) => column.type === CellType.SOURCE_FILE
-			);
-			if (sourceFileColumn) {
-				rows.forEach((row) => {
-					const { cells } = row;
-					const cell = cells.find(
-						(cell) => cell.columnId === sourceFileColumn.id
-					);
-					if (!cell)
-						throw new CellNotFoundError({
-							columnId: sourceFileColumn.id,
-						});
-					const { content } = cell;
-					const frontmatter =
-						app.metadataCache.getCache(content)?.frontmatter;
-					if (frontmatter) {
-						for (const key of Object.keys(frontmatter)) {
-							const value = frontmatter[key];
-
-							let type: FrontMatterType = "text";
-							if (key === "tags" || Array.isArray(value)) {
-								type = "array";
-							} else if (typeof value === "number") {
-								type = "number";
-							} else if (typeof value === "boolean") {
-								type = "boolean";
-							} else {
-								type = "text";
-							}
-
-							const existingKeys =
-								newFrontmatterKeys.get(type) ?? [];
-							const newKeys = [...existingKeys, key];
-							const newKeysSet = new Set(newKeys);
-							newFrontmatterKeys.set(
-								type,
-								Array.from(newKeysSet)
-							);
-						}
-					}
-				});
-			} else {
-				newFrontmatterKeys.clear();
-			}
-			return newFrontmatterKeys;
-		});
-	}, [setFrontMatterKeys, loomState]);
+		setFrontMatterKeys(() =>
+			deserializeFrontmatterKeys(app, columns, rows)
+		);
+	}, [setFrontMatterKeys, loomState, app]);
 
 	function handleSourceAdd(type: SourceType, name: string) {
 		logger("handleSourceAdd");
