+++ src/react/loom-app/app/hooks/use-source/types.ts
@@ -1,5 +1,3 @@
 import { SourceType } from "src/shared/loom-state/types/loom-state";
 
 export type SourceAddHandler = (type: SourceType, name: string) => void;
-
-export type FrontMatterType = "text" | "number" | "boolean" | "array" | "tags";
