+++ src/shared/deserialize-frontmatter/types.ts
@@ -1,0 +1,1 @@
+export type FrontMatterType = "text" | "number" | "boolean" | "array" | "tags";
